Add e2e test
